requirements.txt
Flask
flask-wtf
flask-limiter
python-dotenv
argon2-cffi
email-validator
Flask-Session
redis
pymongo
requests